+++ go-src/entity/gift.go
@@ -22,7 +22,7 @@ type LiveGift struct {
 	Name        string     `gorm:"size:64;comment:礼物名称" json:"name"`
 	Icon        string     `gorm:"size:255;default:'';comment:图标URL" json:"icon"`
 	Animation   string     `gorm:"size:255;default:'';comment:动画资源URL" json:"animation"`
-	Price       uint64     `gorm:"default:0;comment:钻石单价" json:"price"`
+	Price       float64    `gorm:"default:0;comment:钻石单价" json:"price"`
 	Category    string     `gorm:"size:32;default:'';comment:分类" json:"category"`
 	Sort        int        `gorm:"default:0;comment:排序值(越大越靠前)" json:"sort"`
 	Status      uint8      `gorm:"default:0;comment:状态(0-下架,1-上架)" json:"status"`
